Move namespace permission queries behind db traits
